directed precedence graph!
using a similar model to the TypeTable, precedences now use a handle (a
wrapped uint for type checking) and a directed graph represented using
non-exhaustive adjacency lists. as precedences are defined, it's able to
detect circular definitions. now precedences can be defined at runtime
and out of order, and then compared.