Rule revision + enough flexibility for ternary expressions!
I have no desire to actually include ternary expressions as a core
feature in fungus, but having a parser that works well enough for them
to function is INSANE

major addition here is swapping out a predetermined ty + cty pair in
Rules for rule hooks which take in a collapsable expr and provide the
right type checking + type info. this feels very flexible and nice!
rules are also only getting simpler to define